backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.4.2
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
email-validator==2.3.0
reportlab==4.0.7
weasyprint==60.2
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.1
pandas==2.1.3
openpyxl==3.1.2
python-dateutil==2.8.2
pytz==2024.1
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0